refactor(events_api): reorganize event files
